Fix imports; refactor `add_handlers` into `handler.py`
