+++ plugins/post.py
@@ -4,9 +4,8 @@ from bs4 import BeautifulSoup
 from telegram import Update, ReplyKeyboardMarkup, ReplyKeyboardRemove, InputMediaPhoto
 from telegram.constants import ParseMode
 from telegram.error import BadRequest
-from telegram.ext import CallbackContext, ConversationHandler
+from telegram.ext import CallbackContext, ConversationHandler, CommandHandler, MessageHandler, filters
 from telegram.helpers import escape_markdown
-from telegram.ext import CommandHandler, MessageHandler, filters, ConversationHandler
 
 from config import config
 from logger import Log
